requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2
ruff==0.8.6
squareup==43.1.0.20260318